fix(runtime): route spawn_blocking shortcut through Layer 3 Executor
The `runtime/async/shortcuts.rs::spawn_blocking` shortcut called the
primitive `async_utils::spawn_blocking` directly, bypassing the Layer 3
builder and diverging from its sibling `spawn()`, which delegates to
`Executor::new().spawn(future)`. The root cause was that the Layer 3
`Executor` struct exposed `spawn()` but not `spawn_blocking()`.

Add `Executor::spawn_blocking`, carrying the observe instrumentation and
the `SPAWN_BLOCKING_STATS` bookkeeping (relocated from shortcuts.rs so
spawn_blocking observability is co-located with the Executor, mirroring
`spawn`). The shortcut is now a thin `Executor::new().spawn_blocking(f)`.
Public API is unchanged: `SpawnBlockingStatistics` and
`spawn_blocking_stats` are re-exported from `executor::` instead of
`shortcuts::`.

Closes #399
